refactor: move decision-local SystemParam bundles to decision/mod.rs
NpcDataQueries, DecisionNpcState, DecisionExtras were defined in
behavior.rs but used exclusively by decision_system. Move them to
decision/mod.rs where they belong. Remove two unused fields
(stealer_q, has_energy_q) from NpcDataQueries that never had reads.
