+++ rust/src/resources.rs
@@ -2424,13 +2424,24 @@ mod tests {
 
         assert!(cache.should_log(5), "selected slot must return true");
         assert!(!cache.should_log(3), "non-selected slot must return false");
-        assert!(!cache.should_log(0), "slot 0 must return false when not selected");
+        assert!(
+            !cache.should_log(0),
+            "slot 0 must return false when not selected"
+        );
 
         // should_log must agree with push: push must store only for selected slot
         cache.push(5, 1, 0, 0, "msg-selected");
         cache.push(3, 1, 0, 0, "msg-other");
-        assert_eq!(cache.logs[5].len(), 1, "selected slot must have a log entry");
-        assert_eq!(cache.logs[3].len(), 0, "non-selected slot must have no log entry");
+        assert_eq!(
+            cache.logs[5].len(),
+            1,
+            "selected slot must have a log entry"
+        );
+        assert_eq!(
+            cache.logs[3].len(),
+            0,
+            "non-selected slot must have no log entry"
+        );
     }
 
     #[test]
@@ -2442,7 +2453,10 @@ mod tests {
         cache.set_slot_faction(20, 2);
 
         assert!(cache.should_log(10), "same faction slot must return true");
-        assert!(!cache.should_log(20), "enemy faction slot must return false");
+        assert!(
+            !cache.should_log(20),
+            "enemy faction slot must return false"
+        );
     }
 
     #[test]
